Codigo Visao Inicial

+++ Visao/Visao/ControladorPID.h
@@ -18,11 +18,11 @@ private:
 public:
 	ControladorPID();
 	ControladorPID(double Kp, double Ki, double Kd, Robo *robo);
-	void PID2(double xb, double yb);
+	void PID2(float xb, float yb);
 	void resetConst(double Kp, double Ki, double Kd);
 	void setConst(double kp, double ki, double kd);
 	void execute(double anguloAtual, double anguloAlvo, double velocidadeLinear, int *pwm);
-	void PID(double xb, double yb);
+	void PID(float xb, float yb);
 	void pidTiaguera(pt::Point alvo, pt::Point atual, double orientacaoAtual);
 	~ControladorPID(void);
 };
